Use tuple for array types if needed
Previously if a msgpack array was decoded without type information, we'd
default to using a `list`. This meant that messages that used `tuple`
for keys would error on decode without type information. We now
automatically use `tuple` by default when necessary for decoding an
`array` type (e.g. when the value is part of a key in a `dict` or
`set`). This allows us to roundtrip more messages successfully without
type information, while still relying on standard msgpack types.