Add `ext_hook` callback to decoder
This takes an optional callable of the form:

```python
def ext_hook(code: int, data: memoryview) -> Any:
    pass
```

If provided, this will be used to decode any extension types found in a
message (except those that are marked via type-annotation to remain as
`Ext` objects). Note that `data` is a memoryview into the larger message
buffer, minimizing data copies. To prevent the whole message buffer from
persisting in memory for too long, the user should ensure they don't
keep a reference to the `data` buffer in the decoded object.

Also renames `ExtType` to `Ext`.