Optimize buffer usage in `encode`/`encode_into` (#37)
This applies a few buffer-oriented optimizations to speedup encoding (by
up to a factor of 2x).

- Move resizing of write buffer to a separate function, and inline
`mp_write`. This allows gcc to do a better job of optimizing small
memcpy operations with little code bloat.

- When returning a `bytes` object from `encode`, don't hit realloc to
shrink the buffer down to the exact output size. Instead, we just shrink
the view presented to Python. This requires a bit of mucking in the
internals of the bytes object, but nothing too onerous. Since the output
of `encode` will usually be immediately written to a socket/file and
then dropped, this shouldn't result in an increase in application memory
usage.