Speedup decoding of Enum types
This optimizes the decoding of Enums, following the same pattern
recently implemented for `IntEnum` types. A custom hashtable object is
generated once per Enum object and stored on the class as the
`__msgspec_lookup__` attribute. The hashtable maps each enum name to its
value, and allows us to lookup the enum values without allocating a
temporary string object and using it to lookup the enum value.

Using this method, I get a 10x speedup on decoding enums. This
implementation will also be used in a follow-up PR for implementing
decoding of `Literal` string types.