Micro-optimize `decode(type=Struct)`
For the common case of `msgspec.json.decode(msg, type=Struct)`, we can
allocate the `TypeNode` on the stack and have a faster init path,
reducing overhead per call. This still isn't as fast as creating a
`Decoder` once and using `Decoder.decode`, but it brings things closer.
On my machine, decoding a simple `Point(x=1, y=2)` struct takes the
following times:

- `decode` (master): 240 ns
- `decode` (this PR): 150 ns
- `Decoder.decode`: 120 ns