Reduce module state lookup in hot paths
Looking up the module state is generally pretty quick, but if done in a
hot path (e.g. encoding/decoding when not raising an exception) it can
add up. We now store a reference to the module state on the encoder,
which reduces the cost of serializing `enum.Enum` objects by 20-30%, and
negligibly improves a few other encode paths. We also micro-optimize a
few other places where we access the module state. This could be done on
the decoder as well, but (almost) all access in the decoder is for
raising errors and so wouldn't improve performance in common hot paths.