Fixup decode(..., type=Struct) optimization
Between the original PR being made and merged a change went in that
broke the optimization (specifically, the type code for struct types was
split into `MS_TYPE_STRUCT` and `MS_TYPE_STRUCT_ARRAY`). This fixes the
bug, and expands the tests to provide a more localized check for this
behavior.