Add support for renaming fields
Adds a `rename` option to a struct definition. This takes either a
string describing a rename method (``"lower"``, ``"upper"``,
``"camel"``, or ``"pascal"`` for lowercase, UPPERCASE, camelCase, or
PascalCase respectively), or a callable from str -> str that will be
called on every field name. Renamed fields are used for
encoding/decoding only, and are intended for cases where you need the
JSON field names to be something different than what you wish to use in
Python (a common case would be applying camelCase to encoded JSON, but
using snake_case in python).