Fix ARM builds
Some ARM compatibility fixes. The main issue here was that C leaves it
up to the implementation whether `char` is a signed or unsigned type. In
x86 systems it's generally a signed type, but on linux ARM it's
unsigned. OSX ARM defines it as a signed type for compatibility with x86
OSX.