Speedup struct decoding
This further optimizes the decoding of Struct types by rearranging the
field lookup loop a bit to avoid `%` operations. This optimization
is most effective on gcc, resulting in a 20% speedup in decode time.
Clang already did most of this optimization automatically. With this
change, on gcc `bench.py` decreases from ~700 us to ~550 us for json
decoding.