Rename `nogc` to `gc` (#124)
All other option names are `_` separated, so this should really be
`no_gc`. The double-negative feels unnecessary, it's cleaner just to
name this `gc`, defaulting to `gc=True`.

This is a breaking change, but afaik no one is really using this option
right now so a breaking change feels fine.