Add `ValidationError`
Adds a new `ValidationError` error class. This is a subclass of
`DecodeError`, and is used when the decoded message is syntactically
valid but mistyped. `DecodeError` is still used for all other errors
during decoding.

This makes it easier to specialize error handling to differentiate
between syntactically valid (but mistyped) JSON documents and JSON
syntax errors.