Move benchmark plots to vega-lite
This changes the benchmark plots to use `vega-lite` rather than `bokeh`.
This change is nice for a few reasons:

- No need for another Python dependency, plotting code is handled by JS
in the docs source alone.
- The delayed loading of the plots no longer messes with anchor links.
Previously links to a section in the benchmark docs would be off since
the plot rendering would cause the page to scroll after loading. There's
likely a way to avoid this with `bokeh` too, but I couldn't figure it
out.
- The plots are more readable on mobile. Bokeh's responsive plots would
cause the plot labels to squish together, becoming unreadable. The
legend being outside the plot boundary also helps prevent crowding.

This also lets us easily export plots as SVGs, which makes inlining them
into the README easier.