Support mappings for `rename`
This adds support for passing either a `Mapping[str, str]` or a callable
of signature `Callable[[str], Optional[str]]` (the `Optional` is the new
bit) to `rename`. This makes it easier to only rename a few fields in a
struct, as the rename function doesn't need to handle every field.