Add conda-repodata example
This example benchmarks a few different JSON libraries, and compares
using them to parse and query the `current_repodata.json` file from
conda-forge. It's a good example of using `Struct` types to define
smaller "views" of the data only including the needed fields for an
operation. This reduces allocations, significantly increasing
performance.