Fix flaky NamedTuple/TypedDict tests
The tests that a JSON incompatible `NamedTuple`/`TypedDict` would raise
a nice error message previously would sometimes fail with a
`SystemError` indicating a `NULL` return value without a valid
exception. This was tricky to track down due to how rare the failure was
(~1/500 runs).

The issue turns out to be that `PyObject_DelAttr` may sometimes clear an
existing exception, even if no new exception is raised. Why this happens
isn't clear to me; it may be a bug in CPython, or expected behavior.

To fix this, we now use `PyErr_Fetch`/`PyErr_Restore` to ensure the
proper exception persists through the `PyObject_DelAttr` calls.