Add `extra` field to `msgspec.Meta`
This adds a new `extra` field to `msgspec.Meta` annotations. This field
takes a dict of user-defined additional metadata. For the most part
``msgspec`` will ignore this dict entirely (leaving it up to the user
how it's used). The one exception is the tooling in `msgspec.inspect`,
which will extract it and present it as part of a
`msgspec.inspect.Metadata` type node.