Inline common msgpack encode dispatch
This selectively inlines and reorganizes code in the `mpack_encode`
dispatch, following the same optimization already applied to the JSON
encoder.

On my machine this results in up to a 15% speedup in encoding speed.

Inlining is done selectively, only in collections where the value types
are likely to be homogenous (resulting in better branch prediction). The
increase in library size is tiny (136 bytes), and no regressions were
found in our benchmark suite.