Move readme to markdown
The recent changes to freshen up the readme don't render properly on
PyPI, because PyPI doesn't support

- The `raw` restructured text directive (but they do support raw html in
  markdown docs).
- The html `<picture>` tag in general

To work around this we move the readme to markdown, and drop the
light/dark rendering of the logo in the GitHub readme. We could support
this by transforming the readme before uploading to PyPI, but that seems
like too much work for too little benefit.