Add option to define `__dict__` on `Struct` type
This adds a new `dict` option on a `Struct` type, disabled by default.
Enabling this will add a lazily-allocated `__dict__` attribute to all
instances of that type, which can be used to store additional runtime
state. Note that any data stored in the `__dict__` is _not_
encoded/decoded/pickled/copied by default.

One use of this is to support `functools.cached_property` on `Struct`
instances, since this decorator stores the cached value on the instance
`__dict__`.