Fix integer `lt` constraint bug
There was a bug in the handling of integer `lt` constraints that would
only show up if that constraint was also mixed with a lower bound
(`gt`/`ge`) constraint.

```python
Annotated[int, Meta(gt=0, lt=10)]
```

The bug is specific to `lt`, and doesn't appear when using `le`. A
workaround to use for past releases is to use `le` instead of `lt` for
integer lower bounds.

```python
Annotated[int, Meta(gt=0, le=11)]
```