Use `NODEFAULT` instead of `UNSET`
This makes the internal `NODEFAULT` singleton public, and switches the
all previous usage of `msgspec.UNSET` to use `msgspec.NODEFAULT`
instead.

This is a *breaking change*. In the unlikely case a user was using
`msgspec.UNSET` directly (it's a fairly fringe API to use), they should
move to using `msgspec.NODEFAULT` instead. This breaking change is
needed so we can reclaim the `msgspec.UNSET` singleton for an
alternate purpose.