Use enums for singletons in type stubs
This lets type checkers do a better job of type narrowing. For ease of
implementation we only make this change in the type stubs. In normal
usage the singleton object and types should be treated as opaque objects
by users, lieing in the type stubs here seems unlikely to cause an
issue.