Support `typing.Final` annotations
This adds support for `typing.Final` annotations. `Final` can be used to
wrap an existing field annotation, marking it as a field that can't be
modified once it's initialized. This has the same semantics as
`frozen=True`, but only for a single field, and not enforced at runtime.

The `Frozen` annotation may be used to wrap field annotations on any
object-like type we support (`msgspec.Struct`, `attrs`, or
`dataclasses`).