Use `setup.py develop` for now
Some change in `setuptools` or `pip` seems to have dropped the ability
to retain the build directory, which we need for generating coverage of
the .c files. I've spent a bit of time trying to make this work with
`pip install -e .` and can't seem to get things to work. We'll use the
deprecated command for now.