Passthrough input unchanged for `convert` to `Any` type
Previously when converting a value to an `Any` type, we'd recurse
through the input if the input was a known builtin type. We now return
the input unchanged when converting to an `Any` field. This is
significantly more efficient in these cases, as it doesn't require
recursing through the input data.