Support encoding decimals as numbers
Previously we always encoded `decimal.Decimal` values as strings. While
this is the best format for these types (prevents precision loss, widely
supported), sometimes there is a need to encode `Decimal` objects the
same as numbers. This adds a new `decimal_format` option to all
`Encoder` classes. This value defaults to `"string"` (to encode decimals
as strings), but may be set to `"number"` to encode them the same as the
protocol's numeric types.

We opt not to support this configuration in the top-level `encode`
functions for now, since this is a less-common setting.