Add `msgspec.json.Encoder.encode_lines`
This adds a new method on the JSON `Encoder` class for encoding an
iterable of items as newline delimited JSON, one item per line. This is
roughly equivalent to (but up to 3x faster than):

```
def encode_lines(items):
    return b''.join(
        msgspec.json.encode(item) + b'\n' for item in items
    )
```

We limit it to an implementation on the `Encoder` class alone for now
(rather than a new top-level function) since this is a relatively niche
feature.