Support `__post_init__` for `Struct` types
This adds support for an optional `__post_init__` hook method on
`msgspec.Struct` types. If defined, this method is called at the end of
the generated `__init__` method. It can be used to customize the builtin
`__init__` as needed, as well as provides a spot to do some additional
validation.

As with `dec_hook` errors, any `ValueError` or `TypeError` raised in
`__post_init__` will be wrapped in a `ValidationError` when decoding.